coin_auto_trading/requirements.txt
# 코어 의존성
numpy>=1.24.0
pandas>=2.0.0

# 데이터베이스
# SQLite는 Python 표준 라이브러리

# API 클라이언트 (Upbit)
pyupbit>=0.2.28
requests>=2.31.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 테스팅
pytest>=7.4.0
pytest-cov>=4.1.0

# 로깅 (표준 라이브러리 사용하지만 필요시)
# colorlog>=6.7.0

# 타입 힌팅
typing-extensions>=4.8.0

